:sparkles: Add run command to mcp
